requirements.txt
tensorflow==2.7
pillow
biopython
fastapi
uvicorn
python-multipart
sqlalchemy